image_classifier moved under src/
