+++ pyproject.toml
@@ -14,7 +14,7 @@ python_version = "3.10"
 ignore_names = ["forward*"]
 make_whitelist = true
 min_confidence = 60
-paths = ["image_classifier"]
+paths = ["src/image_classifier"]
 sort_by_size = true
 
 #### Python packaging ####
