+++ src/image_classifier/main.py
@@ -8,9 +8,9 @@ import torch.nn as nn
 import torch.optim as optim
 
 from image_classifier.data.data import get_db_connect_object, get_train_test_data_loader
-from image_classifier.model.model import get_model, save_model
-from image_classifier.model.training import train
-from image_classifier.model.validation import validation
+from image_classifier.modeling.model import get_model, save_model
+from image_classifier.modeling.training import train
+from image_classifier.modeling.validation import validation
 from image_classifier.visualization.plot import save_metrics_plot
 
 # Setting the logging level of other loggers to WARNING
